+++ base/cvd/cuttlefish/host/commands/cvd/fetch/BUILD.bazel
@@ -106,6 +106,7 @@ cc_library(
         "//cuttlefish/host/commands/cvd/fetch:fetch_cvd_parser",
         "//cuttlefish/host/commands/cvd/fetch:fetch_tracer",
         "//cuttlefish/host/commands/cvd/fetch:substitute",
+        "//cuttlefish/host/commands/cvd/fetch:vector_flags",
         "//cuttlefish/host/libs/config:fetcher_config",
         "//cuttlefish/host/libs/image_aggregator",
         "//cuttlefish/host/libs/web/http_client:curl_global_init",
@@ -136,8 +137,7 @@ cc_library(
         "//cuttlefish/common/libs/utils:flag_parser",
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/host/commands/cvd/fetch:build_api_flags",
-        "//cuttlefish/host/libs/web:android_build_string",
-        "//cuttlefish/host/libs/web:chrome_os_build_string",
+        "//cuttlefish/host/commands/cvd/fetch:vector_flags",
         "//libbase",
     ],
 )
@@ -209,5 +209,24 @@ cc_library(
 clang_tidy_test(
     name = "substitute_clang_tidy",
     srcs = [":substitute"],
+    tags = ["clang_tidy", "clang-tidy"],
+)
+
+cc_library(
+    name = "vector_flags",
+    srcs = ["vector_flags.cc"],
+    hdrs = ["vector_flags.h"],
+    copts = COPTS,
+    deps = [
+        "//cuttlefish/common/libs/utils:flag_parser",
+        "//cuttlefish/common/libs/utils:result",
+        "//cuttlefish/host/libs/web:android_build_string",
+        "//cuttlefish/host/libs/web:chrome_os_build_string",
+    ],
+)
+
+clang_tidy_test(
+    name = "vector_flags_clang_tidy",
+    srcs = [":vector_flags"],
     tags = ["clang_tidy", "clang-tidy"],
 )
